Consolidate RevitProjectCloseLogger into single file
