+++ RevitProjectCloseLogger/RevitProjectCloseLogger.cs
@@ -1,11 +1,15 @@
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Text;
+using System.Text.Json;
+using Autodesk.Revit.ApplicationServices;
+using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.DB.Events;
 using Autodesk.Revit.UI;
-using Autodesk.Revit.ApplicationServices;
 
 namespace RevitProjectCloseLogger
 {
@@ -29,14 +33,11 @@ namespace RevitProjectCloseLogger
                 ctrl.DocumentClosing += OnDocumentClosing;
                 ctrl.DocumentSynchronizedWithCentral += OnDocumentSynchronizedWithCentral;
 
-                // Compute and cache Revit version info
                 RevitVersionInfo = $"{ctrl.VersionName} {ctrl.VersionNumber} {ctrl.SubVersionNumber}";
 
-                // Create Ribbon UI
                 try
                 {
-                    // Create or get tab
-                    try { application.CreateRibbonTab(RibbonTabName); } catch { /* tab may already exist */ }
+                    try { application.CreateRibbonTab(RibbonTabName); } catch { }
                     var panel = GetOrCreatePanel(application, RibbonTabName, RibbonPanelName);
 
                     var asmPath = System.Reflection.Assembly.GetExecutingAssembly().Location;
@@ -101,10 +102,7 @@ namespace RevitProjectCloseLogger
                     DocumentOpenTimes[e.Document] = DateTime.Now;
                 }
             }
-            catch
-            {
-                // ignore
-            }
+            catch { }
         }
 
         private static void OnDocumentSynchronizedWithCentral(object sender, DocumentSynchronizedWithCentralEventArgs e)
@@ -119,10 +117,7 @@ namespace RevitProjectCloseLogger
                 }
                 DocumentSyncCounts[doc]++;
             }
-            catch
-            {
-                // ignore
-            }
+            catch { }
         }
 
         private static void OnDocumentClosing(object sender, DocumentClosingEventArgs e)
@@ -136,7 +131,6 @@ namespace RevitProjectCloseLogger
 
                 var now = DateTime.Now;
 
-                // Build log data
                 var data = new LogData
                 {
                     UserId = GetUserId(doc.Application),
@@ -173,11 +167,10 @@ namespace RevitProjectCloseLogger
             }
             catch (Exception ex)
             {
-                try { TaskDialog.Show("Project Close Logger", $"Logging failed: {ex.Message}"); } catch { /* ignore UI failures */ }
+                try { TaskDialog.Show("Project Close Logger", $"Logging failed: {ex.Message}"); } catch { }
             }
             finally
             {
-                // Cleanup per-document maps to avoid leaks
                 try { if (DocumentOpenTimes.ContainsKey(e.Document)) DocumentOpenTimes.Remove(e.Document); } catch { }
                 try { if (DocumentSyncCounts.ContainsKey(e.Document)) DocumentSyncCounts.Remove(e.Document); } catch { }
             }
@@ -187,7 +180,6 @@ namespace RevitProjectCloseLogger
         {
             try
             {
-                // Revit username if set; fallback to OS username
                 var revitUser = app.Username;
                 if (!string.IsNullOrWhiteSpace(revitUser)) return revitUser;
             }
@@ -253,10 +245,7 @@ namespace RevitProjectCloseLogger
 
         private static int SafeCount(Document doc, FilteredElementCollector collector)
         {
-            try
-            {
-                return collector.WhereElementIsNotElementType().GetElementCount();
-            }
+            try { return collector.WhereElementIsNotElementType().GetElementCount(); }
             catch { return 0; }
         }
 
@@ -328,6 +317,188 @@ namespace RevitProjectCloseLogger
                 return path.Contains("accdocs") || path.Contains("autodesk docs") || path.Contains("bim 360");
             }
             catch { return false; }
+        }
+    }
+
+    [Transaction(TransactionMode.Manual)]
+    public class ToggleExportCommand : IExternalCommand
+    {
+        public Result Execute(ExternalCommandData commandData, ref string message, ElementSet elements)
+        {
+            try
+            {
+                bool enabled = SettingsManager.IsExportEnabled();
+                bool newValue = !enabled;
+                SettingsManager.SetExportEnabled(newValue);
+
+                var td = new TaskDialog("Project Close Logger")
+                {
+                    MainInstruction = newValue ? "Export ENABLED" : "Export DISABLED",
+                    MainContent = "This setting controls whether a row is appended to the Excel-compatible CSV when a project is closed.",
+                    CommonButtons = TaskDialogCommonButtons.Close
+                };
+                td.Show();
+
+                return Result.Succeeded;
+            }
+            catch (Exception ex)
+            {
+                message = ex.Message;
+                return Result.Failed;
+            }
+        }
+    }
+
+    internal static class SettingsManager
+    {
+        private const string SettingsFileName = "settings.json";
+        private const string AppFolderName = "RevitProjectCloseLogger";
+
+        private class Settings
+        {
+            public bool ExportEnabled { get; set; } = true; // default ON
+        }
+
+        private static string GetAppFolder()
+        {
+            var appData = Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData);
+            var path = Path.Combine(appData, AppFolderName);
+            Directory.CreateDirectory(path);
+            return path;
+        }
+
+        private static string GetSettingsPath()
+        {
+            return Path.Combine(GetAppFolder(), SettingsFileName);
+        }
+
+        public static bool IsExportEnabled()
+        {
+            try
+            {
+                var path = GetSettingsPath();
+                if (!File.Exists(path)) return true; // default enabled
+                var json = File.ReadAllText(path, Encoding.UTF8);
+                var settings = JsonSerializer.Deserialize<Settings>(json);
+                return settings?.ExportEnabled ?? true;
+            }
+            catch { return true; }
+        }
+
+        public static void SetExportEnabled(bool enabled)
+        {
+            try
+            {
+                var settings = new Settings { ExportEnabled = enabled };
+                var json = JsonSerializer.Serialize(settings, new JsonSerializerOptions { WriteIndented = true });
+                File.WriteAllText(GetSettingsPath(), json, Encoding.UTF8);
+            }
+            catch { }
+        }
+
+        public static string GetLogsFolder()
+        {
+            var docs = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            var folder = Path.Combine(docs, AppFolderName, "Logs");
+            Directory.CreateDirectory(folder);
+            return folder;
+        }
+    }
+
+    internal class LogData
+    {
+        public string UserId { get; set; }
+        public string SessionId { get; set; }
+        public DateTime Date { get; set; }
+        public string ProjectName { get; set; }
+        public string ProjectNumber { get; set; }
+        public string ProjectFileName { get; set; }
+        public string Action { get; set; }
+        public DateTime LogStart { get; set; }
+        public DateTime LogEnd { get; set; }
+        public TimeSpan Duration { get; set; }
+        public TimeSpan SessionDuration { get; set; }
+        public int SynchronizedProjectsCount { get; set; }
+        public string UserComputer { get; set; }
+        public string RevitServicePackage { get; set; }
+        public int Warnings { get; set; }
+        public long FileSize { get; set; }
+        public int Worksets { get; set; }
+        public int LinkedModels { get; set; }
+        public int ImportedImages { get; set; }
+        public int Views { get; set; }
+        public int Sheets { get; set; }
+        public int ModelElements { get; set; }
+        public int ModelGroups { get; set; }
+        public int DetailGroups { get; set; }
+        public int DesignOptions { get; set; }
+        public string Diroot { get; set; }
+        public bool DesktopConnector { get; set; }
+        public int ImportedCADFiles { get; set; }
+    }
+
+    internal static class ExcelExporter
+    {
+        private static readonly string[] Header = new[]
+        {
+            "UserId","SessionId","Date","ProjectName","ProjectNumber","ProjectFileName","Action","LogStart","LogEnd","DurationSeconds","SessionDurationSeconds","SynchronizedProjectsCount","UserComputer","RevitServicePackage","Warnings","FileSizeBytes","Worksets","LinkedModels","ImportedImages","Views","Sheets","ModelElements","ModelGroups","DetailGroups","DesignOptions","Diroot","DesktopConnector","ImportedCADFiles"
+        };
+
+        public static void AppendRow(LogData data)
+        {
+            var folder = SettingsManager.GetLogsFolder();
+            var file = Path.Combine(folder, $"RevitLog_{DateTime.Now:yyyyMM}.csv");
+
+            var sb = new StringBuilder();
+
+            if (!File.Exists(file))
+            {
+                sb.AppendLine(string.Join(",", Header));
+            }
+
+            var row = new string[]
+            {
+                Csv(data.UserId),
+                Csv(data.SessionId),
+                Csv(data.Date.ToString("yyyy-MM-dd HH:mm:ss", CultureInfo.InvariantCulture)),
+                Csv(data.ProjectName),
+                Csv(data.ProjectNumber),
+                Csv(data.ProjectFileName),
+                Csv(data.Action),
+                Csv(data.LogStart.ToString("yyyy-MM-dd HH:mm:ss", CultureInfo.InvariantCulture)),
+                Csv(data.LogEnd.ToString("yyyy-MM-dd HH:mm:ss", CultureInfo.InvariantCulture)),
+                Csv(((long)data.Duration.TotalSeconds).ToString(CultureInfo.InvariantCulture)),
+                Csv(((long)data.SessionDuration.TotalSeconds).ToString(CultureInfo.InvariantCulture)),
+                Csv(data.SynchronizedProjectsCount.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.UserComputer),
+                Csv(data.RevitServicePackage),
+                Csv(data.Warnings.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.FileSize.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.Worksets.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.LinkedModels.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.ImportedImages.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.Views.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.Sheets.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.ModelElements.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.ModelGroups.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.DetailGroups.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.DesignOptions.ToString(CultureInfo.InvariantCulture)),
+                Csv(data.Diroot),
+                Csv(data.DesktopConnector ? "1" : "0"),
+                Csv(data.ImportedCADFiles.ToString(CultureInfo.InvariantCulture))
+            };
+
+            sb.AppendLine(string.Join(",", row));
+
+            File.AppendAllText(file, sb.ToString(), new UTF8Encoding(encoderShouldEmitUTF8Identifier: false));
+        }
+
+        private static string Csv(string value)
+        {
+            if (value == null) return string.Empty;
+            var needsQuotes = value.Contains(",") || value.Contains("\"") || value.Contains("\n") || value.Contains("\r");
+            var escaped = value.Replace("\"", "\"\"");
+            return needsQuotes ? $"\"{escaped}\"" : escaped;
         }
     }
